websocket/requirements.txt
# temporary fix for this problem https://github.com/benoitc/gunicorn/pull/2581
https://github.com/benoitc/gunicorn/archive/ff58e0c6da83d5520916bc4cc109a529258d76e1.zip#egg=gunicorn==20.1.0
eventlet
python-socketio
requests
gql[requests]